Reformat homepage, move blogs to BlogCard component
- Only show pagination when needed.
- Move the pagination to bottom.
- Only show new blog in first page.
